+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ISSUES FIXED:
 
+* [58](https://github.com/perfectsense/gyro-azure-provider/issues/58): Differentiate between OS Disk and Data Disks in Azure VMs
+* [53](https://github.com/perfectsense/gyro-azure-provider/issues/53): Gyro does not clean up auto-created VM disks
 * [48](https://github.com/perfectsense/gyro-azure-provider/issues/48): Refactor Sql (QA, ResourceFinder, Direct Dependency, Copyable).
 * [49](https://github.com/perfectsense/gyro-azure-provider/issues/49): Public Ip Address Not Set at Startup for Network Interface.
 * [51](https://github.com/perfectsense/gyro-azure-provider/issues/51): Implement Lifecycle for Azure Storage.